Fix incorrect ChangedValueAtSameVersionException.
I had moved up the setting of this.value so that it would be part of the logged error. However, setting this.value prior to this.lastChangedVersion may lead to a race condition, so exceptions are being thrown when they shouldn't be.

I will attempt to follow up with a threaded test to catch this.

RELNOTES: None.
PiperOrigin-RevId: 219694529