Specify owners of inputs in LostInputs exception
Also adds a getReverseExpansion method to Artifact.ArtifactExpander,
intended for use when creating InputOwners for lost inputs.

Clarifies what Skyframe nodes are in the RewindPlan created by
ActionRewindStrategy, and why.

These are: the node associated with the currently executing (failed)
action, the nodes associated with the actions which create the inputs
which are lost, and all the nodes between them.

The reason: restarting those nodes ensures that reevaluating the current
node will also reevaluate the nodes that will recreate the lost inputs.

RELNOTES: None.
PiperOrigin-RevId: 219696965