workspace_resolved: verify data dependencies
Verify that bazel can resolve data dependencies between repositories
(i.e., one repository needs a file from antoher in its definition)
even if presented in the wrong order in the WORKSPACE file, and also
if read through a resolved file. The latter test gives additional
confidence in our sorting of the repositories in the resolved file,
as the only enforcement (besides reproducibility) made on the order
is respecting chunks, so that loads can succeed.

Related: #6261.

Change-Id: I43aef1f1a3e46ecebe7e89aeca78aad94d3bf992
PiperOrigin-RevId: 220238400