Only report repository return values if containing new information
Repository rules have the option to return a new set of arguments that would
allow them to reproduce identically (whereas the arguments could request
following a branch). Currently, whenever a non-None value is returned, this
is reported to the user; change this to only report, if the return value
contains new information---and in this case, only return what has changed.

While there, also remove the double digesting of the generated directory
in case hashes are verified.

Change-Id: I8bcff45891c755adac72f24c6d07909ed7eb9f3a
PiperOrigin-RevId: 220252511