Improve efficiency of ActionInputMap.
This is currently implemented with open addressing (linear scanning). Instead,
use closed addressing with linked lists in each buckets, but store the pointers
of the linked list in a flat array.

The reason here is that the assignments into an array of a specific type are
expensive (probably doing type checks). Using an Object[] internally is much
more efficient.

That in turn means that we need to store three things (keys, values, and exec
paths). As soon as we need three things, closed addressing is likely more
memory efficient.

This should keep the created garbage about the same while substantially
improving performance.

RELNOTES: None.
PiperOrigin-RevId: 220289296