repository_ctx.download{_and_extract} return a struct.
The methods currently return a struct with an entry for "sha256". If a sha was provided in the rule and it was the correct sha, that sha will be returned. If no sha was provided, the calculated sha will be returned.

Going forward, removing entries from the returned struct will be subject to proper deprecation periods, but additions can happen at any time. Repository rule writers should not depend on the absence of entries in the struct.

RELNOTES[NEW]: repository_ctx.download and repository_ctx.download_and_extract now return a struct.

PiperOrigin-RevId: 220474155