pkg_tar: when passing arguments, quote file names properly
The pkg_tar function passes the list of files to be archived as
arguments (written to an args file). Moreover, the information
which name the file should have in the archive is passed in the
same argument, separated by an equal sign. This obviously causes
problems if the file name itself contains an equal sign. Fix this, by
quoting the file name; to keep the changes to the protocol minimal,
just quote the equal sign by backslash, after escaping bachslashes
by double-backslashes. (Theoretically, it should not be necessary
to keep changes minimal, as both sides of the protocol are embedded
in the same Bazel binary---but better safe than sorry.)

Fixes #4063

...well, kind of, execpt that files are passed as labels and
hence may not contain a lot of characters. (The fact, that newline
characters are not allowed is actually necessary for the concept
of the flagfile to no break.)

Change-Id: I43bab3a0bc8ce26e170cbad3ae64bccb401e3246
PiperOrigin-RevId: 220643705