Fix fileset derived input rewinding
An action that fails because it depends on a fileset that contains an
derived input which was lost will now rewind the appropriate Skyframe
nodes.

Previously rewinding would have failed because the lost input wasn't a
direct dep of the failed action.

Refactors the ActionRewindStrategy logic to make use of the InputOwners
information newly available in the LostInputsExecException, and to be
slightly more lenient for edge cases such as lost intermediates in
multi-Spawn actions.

Also revises RewindPlan's verbiage for actions: "restart" instead of
"reset". Maintaining distinct verbs for distinct types of objects in
ActionRewindStrategy became too cumbersome. However,
SkyframeActionExecutor.resetActionExecution(Action) still uses "reset"
because a reader could reasonably expect a method named
"restartActionExecution" to have actually restarted the action somehow
before it returns, but the method doesn't do that.

RELNOTES: None.
PiperOrigin-RevId: 220653107