Allow repository remapping to work on bzl files loaded, both during workspace evaluation and in BUILD files. Without this, repository remapping is broken.
To enable this, two significant changes were made. The first is to SkylarkImportLookupKey to include the workspace chunk that a particular load originated in to be able to get the remappings. The second is in SkylarkImportLookupFunction to ensure that when we get a load statement that we first do a lookup in the remappings map to get the correct bzl file.

See https://docs.google.com/document/d/1YBltt_jd4QTRzxkp1amsgfS6P8yBVDF4n1KGb3jcf_w/edit# for more information.

Fixes #3115.

RELNOTES: None
PiperOrigin-RevId: 220741830