repository context: support status updates
Add a function to the repository context to update the current status
of the rule execution. As repository rules might take quite some time
to set up an external repository, it is desirable for the user to know
what the rule is currently doing, e.g., fetching, unpacking, patching.

Fixes #1289. Fixes #6427.

Change-Id: I1e812d43d7eba900f266abf70fe00bf1917f198f
PiperOrigin-RevId: 220774485