Support rewinding when inputs of trees are lost
An action that fails because it depends on a derived TreeFileArtifact
which was lost will now rewind the appropriate Skyframe nodes.

Previously, rewinding didn't support lost TreeFileArtifacts which
were indirectly depended on (i.e. when the action directly depended on
the tree artifact, but not directly on the lost file contained by it).

Also includes a couple of drive-by improvements to related comments.

RELNOTES: None.
PiperOrigin-RevId: 220822879