Convert googleapis and remoteapis into local_repository repos.
This is the third_party half of the change to fix #6601.

This half adds the needed BUILD files and WORKSPACE files for
@googleapis and @remoteapis.

The next commit will update the (non-third_party) code to use this.

This commit is very likely to break CI. Instead of rolling back, wait
for the next in the series which should unbreak everything.