Add legacyNamed=true to a number of locations using BuiltinCallable
There is a bug in BuiltinCallable (much as there was in SkylarkSignature) that all parameters were allowed to be named paramters, even if @Param.named was false. A parallel change will fix this bug, but that means we need to manually mark affected parameters so that the fix does not become a breaking change.

RELNOTES: None.
PiperOrigin-RevId: 220841910