Move reads of some env vars to functions.
Slight refactor to the Bash completion script where we move reads of environment variables (whose values we expect users to mutate) to functions expected to be defined in a header file, that is expected to be prepended to the main script file contents. This is a pure refactor with no consequences.

Also, while I'm here, clean up comments.

RELNOTES: None
PiperOrigin-RevId: 221103777