[Remote] Don't discard metadata attached to ExecuteResponse's status.
When executing builds for which input files got removed between
FindMissing() and Execute(), Buildbarn workers will return
FAILED_PRECONDITION errors through ExecuteResponse's status field.
Attached to this status is a violation entry, indicating that the failed
precondition is due to the absence of a blob.

This should cause RemoteModule's RETRIABLE_EXEC_ERRORS to treat this as
a retriable error, causing Bazel to restart the build process for that
action entirely. Unfortunately, it does not, due to the fact that
ExecutionStatusException currently throws away the metadata. Only the
code and message are preserved.

Same idea as https://github.com/bazelbuild/bazel/pull/6641, except does not expose internals, adds a test, and makes sure the Status proto in the Metadata is the same one from which the grpc.Status object is created. Closes #6641.

TESTED=new unit test
RELNOTES: None
PiperOrigin-RevId: 221233439