Windows, test wrapper: implement `tee` for HANDLEs
Implement a GNU coreutils tee(1)-inspired class
for HANDLEs. It copies data from an input HANDLE
to two output HANDLEs.

I will use this logic to redirect the stdout and
stderr of the test process to a log file and to
the test wrapper process' stdout. That is what
test-setup.sh also does.

See https://github.com/bazelbuild/bazel/issues/5508

Change-Id: I1882c1f0f4c1f5375e354849aef80acae73f28bb