bep: close the bep transports in afterCommand
Before this change we'd block an undefined SkyFrame thread when
waiting for the BEP/BES upload to finish. This would also not
allow us to reliably report errors. This change moves the waiting
to afterCommand() where one is allowed to block and can reliably
report errors by throwing an exception.

RELNOTES: None
PiperOrigin-RevId: 221497291