Windows, Bazel client: redirect NUL to stdin
The Bazel client runs the Bazel server with
redirecting NUL to its stdin -- except the file
handle to NUL was not inheritable. Oops.

In this commit:

- make the file handle of NUL inheritable
- update AutoAttributeList to also call
  UpdateProcThreadAttribute

The AutoAttributeList will also be used in the new
test wrapper to redirect (and Tee) the test
process' output. See https://github.com/bazelbuild/bazel/issues/5508

Change-Id: Ib529b381ae6c2a0b705a55838f8f52313f3ffcd5