C++: Adds method for converting CcLinkingInfo
CcLinkParams will go away, instead we will have a new way of propagating linking information through the C++ provider as described in:
https://docs.google.com/document/d/1_irTE1a8aTg8Hhf5mmlEXsFcwOw6-hz3a8FxufCyVKA/edit#heading=h.7795zogmlqtv

Since this will be a huge refactoring, this is being done gradually. In this CL I introduce a function that allows switching between one representation and the other so that work on rules_foreign_cc is unblocked. This gives more time to refactor the rest of the codebase to use the new API.

RELNOTES:none
PiperOrigin-RevId: 222043022