Windows, test wrapper: add CdataEncode function
The test wrapper must output a test XML file.

This XML file contains metadata about the test
(such as its name, duration it ran for, whether it
failed or passed), as well as the test's stdout
and stderr.

Because not every Unicode code point is allowed in
the XML file, and because the test's output is
stored in a CDATA section, we need to escape the
test's output.

This PR adds half of the escaping logic: the part
that replaces invalid octets with question marks.

The function also marks the locations of "]]>" in
the test output, because this string cannot be
embedded in a CDATA section, so a function I'll
add later will replace those in the output.

See https://github.com/bazelbuild/bazel/issues/5508