Post the build tool logs event from the BuildTool
Collect the relevant data in the BuildResult and post the build tool
logs event from a central location rather than a module. Maybe not the
ideal location, but we have multiple modules that need to contribute
to build tool logs, so it must be posted from a common location. It
could be in a module, except that would require modules depending on /
extending other modules, which we don't support right now.

Note that this change includes a fix to LastBuildEvent, without which
this wouldn't work since this change is changing the order of the last
two events posted.

PiperOrigin-RevId: 222061608