C++: Make CcInfo mandatory provider of C++ rules instead of CcSkylarkInfo.
Not every instance in b/77669139 has been removed, however, in this CL we check
for the presence of Java and Python providers to know whether we should give an
error.

This CL introduces a NonCcDepInfo provider to mark these rules that shouldn't be dependencies of C++. This is a workaround that is not exposed to Skylark and should be removed immediately once rules foreign C++ work is unblocked.

RELNOTES:none
PiperOrigin-RevId: 222220381