Flip the flag "--incompatible_static_name_resolution"
This is an incompatible change, but it breaks in a single case:

  global = 2

  def fct(x):
    if x:
      global = 3

    print(global)

Before the change, `global` would refer to either the global variable
or the local variable, depending on the value of `x`. After the change,
it is either `3` or undefined (runtime error).

Fixes #5637

RELNOTES: --incompatible_static_name_resolution is no unable by default
PiperOrigin-RevId: 222242205