Support rewinding of inputs in runfiles
An action that fails because it depends on a generated file in the
action's runfiles will now rewind the appropriate Skyframe nodes.

Previously the checkInputs method lost the relationship between
"middleman" artifacts, used by runfiles and directly depended on by
actions that include runfiles, to the artifacts they aggregate.

This CL refactors checkInputs and the ActionInputMap it populates so
that it can be rerun to reestablish those relationships if the
executing action fails due to lost inputs, without incurring additional
costs in the normal case.

RELNOTES: None.
PiperOrigin-RevId: 222296167