Fix non-determinism in skylark_repository_test.sh.
The test_skylark_repository_which_and_execute test would fail when run alone or when sharding puts it into a different group of tests, because it tried to run "bazel info" before it actually setup its test workspace. This only worked, because the files would have been created (and not cleaned up) by an earlier test.

PiperOrigin-RevId: 222393468