Pass proto compiler to ProtoCompileActionBuilder.
ProtoCompileActionBuilder accepts proto compiler as constructor argument. It can be null. In this case NO_ACTION is created instead throwing NPE. I keep that logic to make this changes no-ops. But I'm not sure that this logic is correct.

Also I was thinking about merge ProtoSourcesProvider protoSourcesProvider and        FilesToRunProvider protoCompilerTarget to something one. But did not find this reasonable to do now.

RELNOTES:none
PiperOrigin-RevId: 222407124