Make shell integration tests compatible with Python 3.
On some systems (e.g. Arch Linux), the Python installed in /usr/bin/python is Python 3.x. This interpreter is used to run certain Python scripts that are part of our shell integration tests. The CL makes them compatible with both Python 2 and 3, so that the tests pass on Arch Linux.

PiperOrigin-RevId: 222426787