Refactor SkyframeActionExecutor
Untangle the execution and completion code paths. The intent is to turn
this into a continuation which calls execute() and then returns. When
the subprocess is done, the Skyfunction is restarted and calls complete()
on the continuation, which performs the rest of the execution for this
action.

This should be a no-op; I'm only doing a minor reordering of steps in the
ActionRunner.

PiperOrigin-RevId: 222531720