Remove warning about an undefined home directory.
If a user doesn't have a home directory, this is usually on purpose (e.g. Blaze runs in a sandbox where the sandboxed user can't access its home directory or doesn't have a home directory or the HOME env var was made to point to /nonexisting). I've seen this warning a lot in sandboxed tests and it's just noise.

I can't think of a case where this would actually catch a legitimate issue.

PiperOrigin-RevId: 222542257