ExecutionTool: close output file of "--explain"
Close the OutputStream of "explanationHandler"
after the handler is uninstalled. This guarantees
that Bazel releases this file promptly, instead of
relying on the GC and finalizers to do so.

Closing streams is important to avoid stream
contention which may prevent deleting files. (Open
files may not be deleted on Windows.)

I (laszlocsomor@) am not currently unaware of any
such contention bugs in this code, but we have
seen intermittent file deletion failures on
Windows, so eliminating a potential source of such
a bug should not hurt.

RELNOTES: none
PiperOrigin-RevId: 222545449