Automated rollback of commit 2bc2efd8bd0af15bc727523a56fdcefe87e2f41d.
*** Reason for rollback ***

b/119861715

*** Original change description ***

bep: close the bep transports in afterCommand

Before this change we'd block an undefined SkyFrame thread when
waiting for the BEP/BES upload to finish. This would also not
allow us to reliably report errors. This change moves the waiting
to afterCommand() where one is allowed to block and can reliably
report errors by throwing an exception.

RELNOTES: None
PiperOrigin-RevId: 222549755