Add an async interface to SpawnAction
Unfortunately, we can't use plain ListenableFuture until all the
implementations have to be updated. Therefore, introduce an intermediate
FutureSpawn abstraction.

Also update the strategy classes to call the new interface, as well as a
bunch of tests.

Only a single Google-internal implementation of SpawnRunner implements
the new interface as of now. Subsequent changes will change the local
and sandboxed spawn runners, as well as the remote spawn runner.

This is part of a larger effort to make action execution asynchronous.
We believe that this will have significant benefits especially for
remote execution: we currently block Skyframe threads waiting for remote
execution to complete, which means we need to increase --jobs to a large
value, which is a significant source of local thread contention.
Additionally, this is at least contributing to the problem of slow
action discovery - the Bazel CLI increases the final action count during
execution as its discovering additional actions. When this effort
completes, actions no longer block Skyframe threads, and those are
therefore free to continue action discovery.

PiperOrigin-RevId: 222606352