C++: Exposes new C++ API to Skylark
This is the API as described in https://docs.google.com/document/d/1_irTE1a8aTg8Hhf5mmlEXsFcwOw6-hz3a8FxufCyVKA/edit

This is a low level API used for rules like cc_import or for foreign C++ rules. This does not include the compilation and linking API elements.

This CL introduces the new API in a compatible way with the old. Several Skylark methods may return different object types depending on whether we are using the new or the old API. We figure out whether we are using the old or the new API from the parameters that are passed to the functions. All of this is transparent to the user from Skylark.

The old API will be deleted once cc_import and foreign rules C++ have been migrated. The internal implementation of CcLinkParams will also be deleted.

RELNOTES:none
PiperOrigin-RevId: 222617935