Remove unused package-loading related code from CppConfiguration
Keeping RedirectChaser there so this is change stays user-invisible. Redirect
chaser ignores visibility, so by removing it builds that used to work can stop
working (if the cc_toolchain should not be visible from cc_toolchain_suite, or
if cc_toolchain was an alias that pointed to somewhere else, and that something
else was not visible).

Observable effect of this cl is that error messages in cases when CROSSTOOL is missing or when --crosstool_top doesn't point to cc_toolchain_suite change. They are moved to the analysis phase, and their wording changed. I expect and hope that this is not going to break anybody.

RELNOTES: None.
PiperOrigin-RevId: 222813249