When using a resolved file, drop implicit parts
If the option `--experimental_resolved_file_instead_of_workspace`
is given, bazel ignores the WORKSPACE file and takes the information
from the resolved file instead. In this case, also ignore the
implicit suffix of every WORKSPACE file, as this information will
be (and, in fact, mostly is already) recorded in the resolved file
anyway. In the long run, we also want to get rid of the implict
prefix, but the resolved file is not there yet.

Change-Id: I9f78251ba67c5464125f01d901b28b3cef1e5b2f
PiperOrigin-RevId: 222947864