Windows, JNI: correctly open stdout/stderr files
Fix the CreateFileW open mode flags
nativeCreateProcess: FILE_APPEND_DATA was useless
without FILE_WRITE_DATA.

Really the intent is to append data to the file,
so we now open the file with GENERIC_WRITE and
seek to its end.

This change will unblock @ulfjack's pending
change [1] about LocalSpawnExecutor where he
intends to use SubprocessBuilder instead of
Command, and which is currently broken on Windows
because tests do not produce any output.

[1] https://bazel-review.googlesource.com/c/bazel/+/82975

Closes #6770.

PiperOrigin-RevId: 222959600