testenv.sh: Write the correct WORKSPACE file in cleanup_workspace.
We create a WORKSPACE file containing a definition for bazel_skylib when first
creating a new workspace directory, but then replaced that WORKSPACE file with
an empty one when cleaning it up.

This fixes a bug where tests would fail when they didn't run as the first test method, because the definition of bazel_skylib would be lost after the first cleanup.

PiperOrigin-RevId: 222978408