Simplify code of getting rules in ExternalPackageUtil.
Since we are reading all the iterable contents in the getRules() function anyway,
and we have the single caller of getRules(), which in fact only returns
a single item wrapped into a list,
we can already return List from selector callback.
This way it becomes easier to chek first found rule.

Also, since "rules" is intended to accumulate the rules,
makes no sence to have it immutable.
Change to ArrayList.

Closes #6775.

PiperOrigin-RevId: 223000317