Remove the entire FileOutErr filtering logic
This is only used on Windows, if parse_showincludes is active. Instead
of filtering on FileOutErr, we ask the SpawnRunner to write the output
to a temporary file, and then post-process that in CppCompileAction.

The previous approach was incompatible with most SpawnRunner
implementations, which do not guarantee that the output is written
through the FileOutErr output streams, but may redirect output to the
underlying files directly. It is a design flaw that FileOutErr creates
this expectation; it would be better to only pass in a pair of Path
objects for the action.

This is in preparation for switching the LocalSpawnRunner to always
redirect output from the local process to a file. This removes the need
for pumping stdout/stderr through the Bazel process (although this was
already not always happening, e.g., when using a process wrapper), and
thereby reducing the need to create two additional threads per
subprocess to perform such pumping.

PiperOrigin-RevId: 223173023