Remove override that made --jobs=auto default to a large value for Blaze.
In unknown commit, we added a hack in Blaze to prevent the new --jobs=auto
value from picking a "too low value" for remote execution when provided
explicitly on the command line.

But arguably that was a mistake. This added a lot of complexity and goes
against our desires to simplify the specification of resource flags, so
remove it. We just rely on an explicit setting in blazerc to cope with
this, and users manually passing --jobs=foo under this scenario should
be aware that they are overriding the default.

RELNOTES: None.
PiperOrigin-RevId: 223198028