Give the xml generating spawn its own out/err
The StandaloneTestStrategy was using the same action execution context
with the same FileOutErr for both the test spawn and the test.xml
generation spawn. If the process wrapper is used, it truncates its
output files to empty files: if the test.xml generation spawn uses
test.log as its stdout, then this causes test.log to become empty.

This is in preparation for enabling --experimental_split_xml_generation
by default.

PiperOrigin-RevId: 223302422