Verify that an action is either shareable or discovers inputs, but not both.
We don't make the guarantee that the execution phase sees only one instance of a shared action and it's therefore possible to get in an inconsistent state when said action is stateful. The simplest way to avoid this is to disallow stateful shared actions.

There were efforts to make this work but they did not work completely and were complicated so they were abandoned.

RELNOTES: None.
PiperOrigin-RevId: 223492814