Begin work on --experimental_better_python_version_mixing
This adds the experimental flag, which will be used to gate the new API and semantics. It also adds the new --python_version flag. The flags are undocumented and incomplete.

The Python version options converter is made more strict to reject non "PY2" and "PY3" values at parse time. We make it an invariant that other values besides PY2/PY3/null are illegal and may result in an exception rather than just a user rule error. (The other option is to split these values from the others into a separate enum, but that seems like a lot of unneeded churn.)

New getter/setter methods are added for retrieving and transitioning the Python mode in a way that respects whether or not the new semantics are enabled. A future CL will reroute existing clients to use these new methods.

Work toward #6583. Design doc for new semantics here: https://github.com/bazelbuild/rules_python/blob/master/proposals/2018-10-25-selecting-between-python-2-and-3.md

RELNOTES: None
PiperOrigin-RevId: 223847968