Create SkylarkCallable.useContext and use it instead of Environment.BazelInfo
This makes the Starlark interpreter pass a StarlarkContext object to methods, which, for Bazel, should be a stand-in replacement for the previous BazelInfo object.

In a future cleanup, we can move Mutability to context, and then turn down useEnvironment. Environment is a bit too powerful an object to be passing around.

RELNOTES: None.
PiperOrigin-RevId: 224216211