Run //src/test/shell/bazel:bazel_bootstrap_distfile_test without network
When testing if we can build a working bazel out of the distribtuion
archive, do so with network access blocked. Packaging a program
requires to be able to build it offline, so verify that we comply
with the standard expectation of how a program is built. Also, build
against the local jdk, as the standard packaging work flow is to
package the build dependencies first and provide them via the build
environment.

Change-Id: I6c044bf941124674bb0b9ef2a2c5b690bed5e0a3
PiperOrigin-RevId: 224342032