TargetCompleteEvent: accept repository targets
Thanks to the "external" virtual package, external repositories appear
in the user visible name space as targets. Therefore, there can also be
TargetConfigured and TargetCompleteEvents for those. Now, as opposed to
build rules, repository rules do not necessarily have an attribute "tags".
Accept this by taking the empty list as list of tags in this case.

Improves on #6725 by fixing the crash in the minimal reproduction example.

Change-Id: I69c5a7906409e15b09e5ae1f6157a02de3482594
PiperOrigin-RevId: 224520041