Fix flakiness in cache stats tests.
Instead of keeping the log of cache stats in a variable and then grepping
for its contents, dump the lines to a file (the test log) and use that as
the target of the grep.  Fixes flakiness observed in a local Bazel run
against this test using 100 iterations, whereas the test would always fail
beforehand at some point.

I don't know why the previous code would flake though.  The string we
grepped showed the right contents so the grep should have always succeeded.
I'm going to leave this at "multiline strings, quoting, and bash oddities"
without investigating further, especially because this change actually
makes the code simpler, which is nicer.

RELNOTES: None.
PiperOrigin-RevId: 225182839