Automated rollback of commit 452fcd0050afa68bf1544492c396806f3395bb7e.
*** Reason for rollback ***

It's ugly and it's not enough: LinkOptions exhibits the same problem and would require the concept of an "action lookup key" to be exposed to Starlark.

*** Original change description ***

Adds the label of the cc compilation context providing the linkstamp to the Linkstamp class. Uses
that for equals instead of reference equals of the NestedSet.

Warning message now emits the offending rules.

Small AutoCodec cleanup.

PiperOrigin-RevId: 225337786