Change static_link_srcs feature to static_link_test_srcs.
Tying this feature only to dynamic mode binaries has unwanted side effects of binaries that are dependencies to tests also being effected by this, leading to build breakages. Tying this feature not only to dynamic mode (which makes sense as we want to take it effect when proto_dynamic_mode_static_link is taking effect as well) but also to the binary being a test solves this issue.

Also defer the retirement of the experimental flag with a similar behavior to next year.

RELNOTES: None.
PiperOrigin-RevId: 225345869