Bazel client: refactor blaze_util::MakeCanonical
The function no longer indicates error by
returning an empty string. Instead it returns a
bool and sets an error message to a user-provided
string pointer.

The benefit of the change is that it's now visible
at call sites whether MakeCanonical's failure is
tolerated or not, and whether empty path is fine
or not. (There was also a TODO for me to do this
refactoring.)

Change-Id: Id34ce561a1254fe091ea315356572952d8c0bd31