block network for //src/test/shell/bazel:workspace_resolved_test
The workspace_resolved_test routinely calls `bazel sync`, as it is an integration
test for a use case starting with this command. Calling `bazel sync` it is easy
to accidentally call out to the network, even though all needed files could be
provided via a distdir. This caused flakyness in the past. So block the network
to notice network calls as they are added instead of later in the form flakyness.

Change-Id: I59f709761215f119cd67ffa62dc50c4d7d3a6476
PiperOrigin-RevId: 225358722