Migrate config_feature_flag to be backed by Starlark build configuration.
This should be a purely internal refactoring, but it means that trimming
that works for Starlark build settings would also work for config_feature_flag.
(The reverse is not true yet. But it's possible that if automatic trimming
turns out to be too far out, manual trimming could be extended to Starlark
build settings.)

RELNOTES: None.
PiperOrigin-RevId: 225464499