Uncomment android_sdk/ndk_repository on Windows
Revive #5023 (by @philwo) so that we can try to revive #4797.

`shell_commands` will put the whole cmd string into an array (`[cmd]`) before passing it to `subprocess.run`. This will make `subprocess.run` thinks that the first element of the array is the full path of the program, hence the weird error in #5023. We work around it with `batch_commands` which pass the cmd string directly into `subprocess.run` and let `subprocess.run` do the heavy-lifting.

/cc @philwo @meteorcloudy

Closes #6699.

PiperOrigin-RevId: 225547132