Make ShowIncludesFilter ignore execroot differences
Fixes https://github.com/bazelbuild/bazel/issues/6847

This change is for making including scanning work on Windows for builds with remote caching or remote execution enabled.

After this change, the ShowIncludesFilter will look for the first `execroot\<workspace_name>` in the output header file paths, then it considers `C:\...\execroot\<workspace_name>` as the execroot path. Because execroot path could be different if remote cache is hit, we ignore it and only add the relative path as dependencies.

I'm quite unwilling to make this change, because parsing `execroot\\<workspace_name>` for execroot is not guaranteed to work always. But considering the only case this could go wrong is when people use an output base that already contains `execroot\\<workspace_name>`, which I think should never happen.

Closes #6931.

Change-Id: Ife2cb91c75f1b5b297851400e672db2b35ff09e0
PiperOrigin-RevId: 225553627