Bazel client: fix reporting of ignored RC files
Compute the difference between the read RC files
and the ignored ones based on their canonical
paths, not on the textual value of the path.

This aids for proper reporting of ignored RC
files, while reporting them with a path familiar
to the user (e.g. paths the user passed with flags
rather than the potentially different canonical
version of the same path).

Also: respect $HOME on Windows when Bazel is
running inside of a test, to avoid picking up the
real RC file from the user's home directory. This
fixes //src/test/cpp:rc_file_test on Windows.

Fixes https://github.com/bazelbuild/bazel/issues/6138

Closes #6910.

PiperOrigin-RevId: 225788973