Handle interrupts (&etc.) before a rewound action re-evaluates
Before rewinding, consumers of ActionStartedEvents and
ActionCompletionEvents could expect them to always arrive in pairs,
regardless of whether a build was interrupted or stopped due to an error.

With rewinding, an ActionCompletionEvent is not emitted for a failed
action that gets rewound.

So that those consumers can maintain their invariants, a failed action
that gets rewound emits an ActionRewoundEvent.

CriticalPathComputer uses ActionRewoundEvent to note that the specified
action is no longer running, but may be restarted.

RELNOTES: None.
PiperOrigin-RevId: 225912544